requirements.txt
torch>=1.12.0
transformers>=4.21.0
biopython>=1.79
numpy>=1.21.0
tqdm>=4.64.0
scikit-learn>=1.0.2
faiss-cpu>=1.7.2
networkx>=2.8.4
matplotlib>=3.5.2
typing-extensions>=4.3.0